COMPLETED! Project is completed in line with the assessment requirements, and not so much what we had planned for in the planning document.
Finished implementation of the Finite State Machine, only issue was that we had not made the functions inside the State class virtual, and so the inheriting state classes were not overriding the function, and a blank function was being run. This of course meant that our duck wasn't moving or responding to the mouse, even though everything was working and compiling perfectly. 

Also wrote the evaluation document for the post production part of the assessment, and made a small comment inside of the planning document. Added a wanderState class to our project, and a bunch of other blank classes should I ever come back and pick this project up again in the future. Though we probably won't.